// src/app/models/compania.model.ts
export class CompaniaModel {
    id:number = 0;
    name: string;
    identification: string;
    picture: string;
    pricturedata: string;
    email: string;
    phone: string;
    webpage: string;
    address1: string;
    address2: string;
    status: string;
    country: string;
    region:string;
    zipcode:string;
}